1st one

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -1,18 +1,18 @@
 package org.launchcode.controllers;
 
+import org.launchcode.models.Category;
 import org.launchcode.models.Cheese;
-import org.launchcode.models.CheeseType;
+
+import org.launchcode.models.data.CategoryDao;
 import org.launchcode.models.data.CheeseDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.List;
 
 /**
  * Created by LaunchCode
@@ -23,6 +23,9 @@ public class CheeseController {
 
     @Autowired
     private CheeseDao cheeseDao;
+
+    @Autowired
+    private CategoryDao categoryDao;
 
     // Request path: /cheese
     @RequestMapping(value = "")
@@ -38,20 +41,51 @@ public class CheeseController {
     public String displayAddCheeseForm(Model model) {
         model.addAttribute("title", "Add Cheese");
         model.addAttribute(new Cheese());
-        model.addAttribute("cheeseTypes", CheeseType.values());
+//        model.addAttribute("cheeseTypes", CheeseType.values());
+        model.addAttribute("categories", categoryDao.findAll());
         return "cheese/add";
     }
 
     @RequestMapping(value = "add", method = RequestMethod.POST)
-    public String processAddCheeseForm(@ModelAttribute  @Valid Cheese newCheese,
-                                       Errors errors, Model model) {
+    public String processAddCheeseForm(
+            @ModelAttribute  @Valid Cheese newCheese,
+            Errors errors, @RequestParam int categoryId,
+            Model model) {
+        Category cat = categoryDao.findOne(categoryId);
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add Cheese");
+            model.addAttribute("categories", categoryDao.findAll());
             return "cheese/add";
         }
-
+        newCheese.setCategory(cat);
         cheeseDao.save(newCheese);
+        return "redirect:";
+    }
+
+    @RequestMapping(value = "edit/{cheeseId}", method = RequestMethod.GET)
+    public String displayCheeseEditForm(Model model, @PathVariable int cheeseId){
+        model.addAttribute("title", "Edit "+ cheeseDao.findOne(cheeseId).getName());
+        model.addAttribute("cheese", cheeseDao.findOne(cheeseId));
+        model.addAttribute("categories", categoryDao.findAll());
+        return "cheese/edit";
+    }
+
+    @RequestMapping(value = "edit", method = RequestMethod.POST)
+    public String editCheese(Model model, @RequestParam int cheeseId, @ModelAttribute @Valid Cheese cheese, Errors errors){
+
+        if (errors.hasErrors()){
+            model.addAttribute("categories", categoryDao.findAll());
+            model.addAttribute("cheeseId", cheeseId);
+            return "cheese/edit";
+        }
+
+        Cheese currCheese = cheeseDao.findOne(cheeseId);
+        currCheese.setName(cheese.getName());
+        currCheese.setDescription(cheese.getDescription());
+        currCheese.setCategory(cheese.getCategory());
+        cheeseDao.save(currCheese);
+
         return "redirect:";
     }
 
@@ -65,11 +99,23 @@ public class CheeseController {
     @RequestMapping(value = "remove", method = RequestMethod.POST)
     public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
 
-        for (int cheeseId : cheeseIds) {
-            cheeseDao.delete(cheeseId);
+        if (cheeseIds != null) {
+            for (int cheeseId : cheeseIds) {
+                cheeseDao.delete(cheeseId);
+            }
         }
 
         return "redirect:";
+    }
+
+    @RequestMapping(value = "category/{id}", method = RequestMethod.GET)
+    public String category(Model model, @PathVariable int id) {
+        Category cat = categoryDao.findOne(id);
+        List<Cheese> cheeses = cat.getCheeses();
+        model.addAttribute("cheeses", cheeses);
+        model.addAttribute("title", "Cheeses in Category: " + cat.getName());
+        model.addAttribute("categoryId", cat.getId());
+        return "cheese/index";
     }
 
 }
